fix: Preview for short urls (#2174)
